Fix mistakes in ORM model & migration code
Working out how to get associations to load revealed various
mistakes in the models/migrations. I changed how products are
created in the seed script to match the documentation. I suspect
the change wasn't necessary, as the source of the bug was elsewhere,
but it works, and I don't feel like testing if the previous way
of saving the data works as well.