Run server with ts-node in prod
It seems that typeorm insists on reaching for the .ts files
even when I'm trying to run a process from dist/*.js, resulting
in an "import outside a module" error.

This feels like the wrong way to run the server in production,
but I just want to get it working and figure out "the correct way"
later.