Remove build step from prod server start
Build is supposed to run in the release phase of deployment, not
the run phase, which might explain why the app keeps crashing.
Also, it seems that config vars in heroku.yml don't do anything,
but documentation is light, so there's not much I can do about
figuring out the problem.