Remove session cart functionality from server
We've decided to handle keeping track of a user's cart on the
client side, since basic cookie sessions don't seem to work as
intended for GQL API calls. This means that we don't currently
need a session or the concept of a cart or cart items on the
server side.