Make sure environment/containers are reset in between tests
We were getting browser test failures in CI that we couldn't
reproduce on local, because the integration test script that
I wrote awhile ago was hot garbage that technically worked but
dirtied the CI environment such that server errors carried over
into the browser tests. The changes to the script files fix this,
but we also call 'docker-compose down' between them to be on the
safe side.