Fix running migrations on release
It seems that Heroku is only willing to run one command, because
the 'echo' showed up in the logs, but the migration didn't. It
seems that the NODE_ENV is getting set correctly, though, so maybe
the problem is with TypeORM. We're a major version behind, though,
and I'm too lazy to look up old versions of the docs to debug.

I made the definition of the DB URL consistent across environments.
Hopefully this all works.