change base images to alpine and remove sha version
The reason we were using buster-slim instead of alpine was due to
a bug with Flow.js, which we haven't been using for awhile.
I removed the specific sha version, because it helped with
image caching when I was using Travis CI, but GitHub actions
is even worse for image caching out-of-the-box, so no need to
bother with being so specific.